Add resize support for logical and physical volumes
The logical volume lwrp was enhanced to support
resizing exisiting logical volumes. If the resize
is set to true in the resource and the lvm
is already existing and smaller than the requested size
it will be extended to the desired size.
The physical volume lwrp is extended with a resize
action which will resize the pv to the current size of
the underlying block device. Intended for use with online
disk extend in virtual environments.